refactor: introduce core systems
