+++ EchoTrail Shared/Core/EntitySystem.swift
@@ -1,54 +1,28 @@
 import SpriteKit
 
-/// Represents any interactive object in the game world.
-/// Entities own their spatial information and a render node that is
-/// eventually attached to the scene graph.
 protocol GameEntity: AnyObject {
-    /// Logical position in the grid.
     var position: IntPoint { get set }
-
-    /// SpriteKit node used for rendering.
     var node: SKNode { get }
-
-    /// Per frame update hook.
-    /// - Parameter deltaTime: Time elapsed since last update.
     func update(deltaTime: TimeInterval)
 }
 
-/// Entities that keep a trail of previous positions and support grid based
-/// movement can conform to `TailEntity` to gain shared behaviour such as
-/// pushing tail segments and attempting moves on the grid.
 protocol TailEntity: GameEntity {
-    /// Previous grid position; useful for movement interpolation.
     var previousPosition: IntPoint { get set }
-
-    /// Ordered list of past positions forming the tail.
     var tail: [IntPoint] { get set }
 }
 
 extension TailEntity {
-    /// Pushes the current position to the head of the tail.
-    /// - Parameter limit: Maximum length of the tail.
     func pushTail(limit: Int = 6) {
         tail.insert(position, at: 0)
         if tail.count > limit { tail.removeLast() }
     }
 
-    /// Attempts to move the entity in the given direction on a grid.
-    ///
-    /// - Parameters:
-    ///   - direction: Direction vector measured in grid units.
-    ///   - passable: Closure determining if a cell can be entered.
-    ///   - pointFor: Maps a grid point to a SpriteKit coordinate.
-    ///   - interval: Animation interval for node movement.
-    /// - Returns: `true` if the entity moved to a new cell.
     @discardableResult
     func tryMove(direction: (Int, Int),
                  passable: (IntPoint) -> Bool,
                  pointFor: (IntPoint) -> CGPoint,
                  interval: TimeInterval) -> Bool {
-        let np = IntPoint(x: position.x + direction.0,
-                          y: position.y + direction.1)
+        let np = IntPoint(x: position.x + direction.0, y: position.y + direction.1)
         previousPosition = position
         if passable(np) {
             position = np
@@ -62,7 +36,6 @@ extension TailEntity {
     }
 }
 
-/// Player controlled entity.
 final class PlayerEntity: TailEntity {
     var position: IntPoint
     var previousPosition: IntPoint
@@ -78,13 +51,11 @@ final class PlayerEntity: TailEntity {
     func update(deltaTime: TimeInterval) { }
 }
 
-/// Follows the player's previous positions with a delay.
 final class EchoEntity: TailEntity {
     var position: IntPoint
     var previousPosition: IntPoint
     var tail: [IntPoint] = []
     let node: SKShapeNode
-    /// Index offset into the player's history that this echo follows.
     var delayIndex: Int
 
     init(position: IntPoint, delayIndex: Int, node: SKShapeNode) {
@@ -97,12 +68,9 @@ final class EchoEntity: TailEntity {
     func update(deltaTime: TimeInterval) { }
 }
 
-/// Static or kinetic obstacle occupying cells on the grid.
 final class ObstacleEntity: GameEntity {
     var position: IntPoint
     let node: SKShapeNode
-
-    /// Path for kinetic obstacles; empty for static ones.
     var path: [IntPoint]
     var currentIndex: Int
     private let pointFor: (IntPoint) -> CGPoint
@@ -136,21 +104,18 @@ final class ObstacleEntity: GameEntity {
     }
 }
 
-/// Central manager responsible for keeping track of all entities and
-/// propagating update calls.
 final class EntitySystem {
     private(set) var entities: [GameEntity] = []
 
-    func add(_ entity: GameEntity, to scene: SKNode) {
+    func add(_ entity: GameEntity, addNode: (SKNode) -> Void) {
         entities.append(entity)
-        scene.addChild(entity.node)
+        addNode(entity.node)
     }
 
     func removeAll() {
         entities.removeAll()
     }
 
-    /// Updates every managed entity.
     func update(deltaTime: TimeInterval) {
         for entity in entities {
             entity.update(deltaTime: deltaTime)
